refactor: extract useRequireAuth, dedup the login-redirect effect (rule of three)
Same 5-line useEffect (!isLoadingSession && !session -> redirect to
/auth/login) was copy-pasted in post/create, post/edit/[id], post/mine,
and now media — 4 occurrences, past the rule-of-three line. Extracted
into src/context/auth/useRequireAuth.tsx (not src/lib/ per the
when-to-create-lib rubric: it depends on next/navigation, so it isn't a
framework-free pure lib) rather than re-exported from context/auth's
index.tsx, to avoid a circular import (useRequireAuth needs useAuth,
which lives in index.tsx).

Side effect: post/edit/[id]'s copy had an incomplete effect dependency
array ([isLoadingSession] only, missing session/router) — the shared
hook uses the correct one, so that inconsistency is gone too.
