+++ docs/.afm-log/events/2026-07.md
@@ -353,3 +353,4 @@
 - [2026-07-26T02:19:29Z] task | s=5d247c | commit
 - [2026-07-26T02:35:34Z] task | s=5d247c | commit
 - [2026-07-26T02:59:55Z] task | s=5d247c | commit
+- [2026-07-26T03:07:22Z] task | s=5d247c | commit
